Moved application tests to e2e folder (update imports).
