+++ sources/packages/backend/apps/api/src/route-controllers/application/_tests_/e2e/application.aest.controller.getCompletedApplicationDetails.e2e-spec.ts
@@ -5,7 +5,7 @@ import {
   BEARER_AUTH_TYPE,
   createTestingAppModule,
   getAESTToken,
-} from "../../../testHelpers";
+} from "../../../../testHelpers";
 import {
   E2EDataSources,
   MSFAAStates,
