+++ sources/packages/backend/apps/api/src/route-controllers/application/_tests_/e2e/application.institutions.controller.getApplicationOverallDetails.e2e-spec.ts
@@ -7,7 +7,7 @@ import {
   getAuthRelatedEntities,
   getInstitutionToken,
   InstitutionTokenTypes,
-} from "../../../testHelpers";
+} from "../../../../testHelpers";
 import {
   createE2EDataSources,
   createFakeInstitutionLocation,
