+++ sources/packages/backend/apps/api/src/route-controllers/application/_tests_/e2e/application.institutions.controller.getInProgressApplicationDetails.e2e-spec.ts
@@ -9,8 +9,8 @@ import {
   INSTITUTION_BC_PUBLIC_ERROR_MESSAGE,
   INSTITUTION_STUDENT_DATA_ACCESS_ERROR_MESSAGE,
   InstitutionTokenTypes,
-} from "../../../testHelpers";
-import { SuccessWaitingStatus } from "../models/application.dto";
+} from "../../../../testHelpers";
+import { SuccessWaitingStatus } from "../../models/application.dto";
 import {
   E2EDataSources,
   createE2EDataSources,
