+++ sources/packages/backend/apps/api/src/route-controllers/application/_tests_/e2e/application.students.controller.applicationCancelChangeRequest.e2e-spec.ts
@@ -7,7 +7,7 @@ import {
   FakeStudentUsersTypes,
   getStudentToken,
   mockUserLoginInfo,
-} from "../../../testHelpers";
+} from "../../../../testHelpers";
 import {
   createE2EDataSources,
   E2EDataSources,
