+++ sources/packages/backend/apps/api/src/route-controllers/application/_tests_/e2e/application.students.controller.getApplicationToRequestAppeal.e2e-spec.ts
@@ -8,7 +8,7 @@ import {
   getStudentToken,
   mockJWTUserInfo,
   resetMockJWTUserInfo,
-} from "../../../testHelpers";
+} from "../../../../testHelpers";
 import {
   saveFakeApplicationDisbursements,
   createE2EDataSources,
