+++ sources/packages/backend/apps/api/src/route-controllers/application/_tests_/e2e/application.students.controller.getApplicationWarnings.e2e-spec.ts
@@ -8,7 +8,7 @@ import {
   FakeStudentUsersTypes,
   getStudentToken,
   mockUserLoginInfo,
-} from "../../../testHelpers";
+} from "../../../../testHelpers";
 import {
   saveFakeApplicationDisbursements,
   createE2EDataSources,
