+++ sources/packages/backend/apps/api/src/route-controllers/application/_tests_/e2e/application.students.controller.getEnrolmentApplicationDetails.e2e-spec.ts
@@ -6,7 +6,7 @@ import {
   FakeStudentUsersTypes,
   getStudentToken,
   getStudentByFakeStudentUserType,
-} from "../../../testHelpers";
+} from "../../../../testHelpers";
 import {
   saveFakeApplicationDisbursements,
   createE2EDataSources,
